LayoutTests/http/tests/webgpu/webgpu/webgpu/idl/exposed.https-expected.txt

CONSOLE MESSAGE: Error: Item/interface should be exposed on secure context

Harness Error (FAIL), message = Error: Item/interface should be exposed on secure context
